@@ -1,24 +1,20 @@
-declare function rodnecislo(rc: string): rodnecislo.RodneCislo;
+export function rodnecislo(rc: string): RodneCislo;
 
-declare namespace rodnecislo {
-  interface RodneCislo extends Object {
-    isValid(): boolean
+export interface RodneCislo {
+  isValid(): boolean
 
-    isMale(): boolean
-    isFemale(): boolean
+  isMale(): boolean
+  isFemale(): boolean
 
-    year(): number
-    month(): number
-    day(): number
+  year(): number
+  month(): number
+  day(): number
 
-    birthDate(): Date
-    birthDateAsString(): string
+  birthDate(): Date
+  birthDateAsString(): string
 
-    isAdult(adulthood?: number): boolean
-    age(): number
+  isAdult(adulthood?: number): boolean
+  age(): number
 
-    dic(): string
-  }
+  dic(): string
 }
-
-export = rodnecislo
